+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 ~~~~~
 - ``RestClient.disseminate`` now accepts ``str`` parameters for ``dip_format`` and ``id_type``
 - ``DIPDownloader.size`` property that indicates the size of the DIP in bytes
+- ``dpres-client dissemination request-dip`` command to request a DIP using the V3 API and optionally download it
 
 Changed
 ~~~~~~~